Поменял тип и проверку на IsEOF в ReadByte

+++ lw/lw3/task3-streams/lib/Common/CompressItem.h
@@ -4,5 +4,5 @@
 struct CompressItem
 {
 	uint8_t byte = 0;
-	size_t count = 0;
+	uint8_t count = 0;
 };
